First round of refactoring.
