+++ estimagic/tests/dashboard/test_utilities.py
@@ -1,5 +1,12 @@
 """Test helper functions for the dashboard."""
+import webbrowser
 from pathlib import Path
+
+import numpy as np
+import pandas as pd
+from bokeh.io import output_file
+from bokeh.io import save
+from bokeh.models import ColumnDataSource
 
 import estimagic.dashboard.utilities as utils
 
@@ -74,3 +81,97 @@ def test_get_color_palette_50():
 
 
 # not testing find_free_port
+
+
+def test_plot_time_series_with_large_initial_values():
+    cds = ColumnDataSource({"y": [2e17, 1e16, 1e5], "x": [1, 2, 3]})
+    title = "Are large initial values shown?"
+    fig = utils.plot_time_series(data=cds, y_keys=["y"], x_name="x", title=title)
+    title = "Test _plot_time_series can handle large initial values."
+    output_file("time_series_initial_value.html", title=title)
+    path = save(obj=fig)
+    webbrowser.open_new_tab("file://" + path)
+
+
+# ====================================================================================
+# map_group_to_params
+# ====================================================================================
+
+
+def test_map_groups_to_params_group_none():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = None
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_nan():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = np.nan
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_empty():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = ["", "", "x", "x"]
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {"x": ["c", "d"]}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_false():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = [False, False, "x", "x"]
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {"x": ["c", "d"]}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_not_none():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = [None, "A", "B", "B"]
+    params.index = ["a", "b", "c", "d"]
+    params["name"] = ["a", "b", "c", "d"]
+    expected = {"A": ["b"], "B": ["c", "d"]}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_int_index():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params.index = ["0", "1", "2", "3"]
+    params["name"] = ["0", "1", "2", "3"]
+    params["group"] = [None, "A", "B", "B"]
+    expected = {"A": ["1"], "B": ["2", "3"]}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_multi_index():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = [None, "A", "B", "B"]
+    params["ind1"] = ["beta", "beta", "cutoff", "cutoff"]
+    params["ind2"] = ["edu", "exp", 1, 2]
+    params.set_index(["ind1", "ind2"], inplace=True)
+    params["name"] = ["beta_edu", "beta_exp", "cutoff_1", "cutoff_2"]
+    expected = {"A": ["beta_exp"], "B": ["cutoff_1", "cutoff_2"]}
+    res = utils.map_groups_to_params(params)
+    assert expected == res
